+++ src/main/java/net/soulsweaponry/items/sword/DarkinBlade.java
@@ -35,7 +35,7 @@ public class DarkinBlade extends UltraHeavyWeapon implements GeoItem {
     private final AnimatableInstanceCache factory = GeckoLibUtil.createInstanceCache(this);
 
     public DarkinBlade(ToolMaterial toolMaterial, Settings settings) {
-        super(toolMaterial, ConfigConstructor.darkin_blade_damage, ConfigConstructor.darkin_blade_attack_speed, settings, true);
+        super(toolMaterial, (int) ConfigConstructor.darkin_blade_damage, ConfigConstructor.darkin_blade_attack_speed, settings, true);
         this.addTooltipAbility(TooltipAbilities.OMNIVAMP, TooltipAbilities.SWORD_SLAM);
     }
 
@@ -46,7 +46,7 @@ public class DarkinBlade extends UltraHeavyWeapon implements GeoItem {
         }
         if (attacker instanceof PlayerEntity player) {
             if (!player.getItemCooldownManager().isCoolingDown(stack.getItem()) && !(player.getHealth() >= player.getMaxHealth())) {
-                this.applyItemCooldown(player, Math.max(ConfigConstructor.lifesteal_item_min_cooldown, ConfigConstructor.lifesteal_item_cooldown - this.getReduceLifeStealCooldownEnchantLevel(stack) * 6));
+                this.applyItemCooldown(player, (int) Math.max(ConfigConstructor.lifesteal_item_min_cooldown, ConfigConstructor.lifesteal_item_cooldown - this.getReduceLifeStealCooldownEnchantLevel(stack) * 6));
                 float healing = ConfigConstructor.lifesteal_item_base_healing;
                 if (ConfigConstructor.lifesteal_item_heal_scales) {
                     healing += MathHelper.ceil(((float)WeaponUtil.getEnchantDamageBonus(stack))/2);
@@ -68,9 +68,9 @@ public class DarkinBlade extends UltraHeavyWeapon implements GeoItem {
                 world.playSound(player, player.getBlockPos(), SoundEvents.ENTITY_BLAZE_SHOOT, SoundCategory.PLAYERS, 1f, 1f);
                 cooldownMod = 0.75f;
                 //NOTE: Ground Smash method is in parent class DetonateGroundItem
-                user.addStatusEffect(new StatusEffectInstance(EffectRegistry.CALCULATED_FALL.get(), 600, ConfigConstructor.darkin_blade_ability_damage));
+                user.addStatusEffect(new StatusEffectInstance(EffectRegistry.CALCULATED_FALL.get(), 600, (int) ConfigConstructor.darkin_blade_ability_damage));
             } else {
-                this.detonateGroundEffect(user, ConfigConstructor.darkin_blade_ability_damage, 0, world, stack);
+                this.detonateGroundEffect(user, (int) ConfigConstructor.darkin_blade_ability_damage, 0, world, stack);
             }
             stack.damage(3, user, (p_220045_0_) -> p_220045_0_.sendToolBreakStatus(user.getActiveHand()));
             this.applyItemCooldown(player, MathHelper.floor(this.getScaledCooldown(stack) * cooldownMod));
@@ -88,8 +88,8 @@ public class DarkinBlade extends UltraHeavyWeapon implements GeoItem {
     }
 
     protected int getScaledCooldown(ItemStack stack) {
-        int base = ConfigConstructor.darkin_blade_ability_cooldown;
-        return Math.max(ConfigConstructor.darkin_blade_ability_min_cooldown, base - this.getReduceCooldownEnchantLevel(stack) * 15);
+        int base = (int) ConfigConstructor.darkin_blade_ability_cooldown;
+        return (int) Math.max(ConfigConstructor.darkin_blade_ability_min_cooldown, base - this.getReduceCooldownEnchantLevel(stack) * 15);
     }
 
     @Override
